move addNewChapter and addNewRecipe to utilities
